+++ _mocks_/fileMocks.js
@@ -1,0 +1,1 @@
+export default 'test-file-stub';
