+++ _mocks_/styleMocks.js
@@ -1,0 +1,1 @@
+module.exports = {};
