+++ package.json
@@ -24,25 +24,25 @@
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.21.2",
     "babel-eslint": "^10.1.0",
-    "css-loader": "^6.7.4",
+    "css-loader": "^6.7.3",
     "csv-loader": "^3.0.5",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5",
-    "hint": "^7.1.8",
-    "html-webpack-plugin": "^5.5.1",
+    "html-webpack-plugin": "^5.5.0",
+    "file-loader": "^6.2.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "sass": "^1.57.1",
     "sass-loader": "^13.2.0",
-    "style-loader": "^3.3.3",
+    "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0",
-    "webpack": "^5.84.1",
-    "webpack-cli": "^5.1.1",
-    "webpack-dev-server": "^4.15.0",
+    "webpack": "^5.75.1",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.0",
     "xml-loader": "^1.2.1"
   },
   "dependencies": {
@@ -50,5 +50,11 @@
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@lhci/cli": "^0.12.0",
     "lodash": "^4.17.21"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+    }
   }
 }
