+++ sum.js
@@ -1,0 +1,4 @@
+function sum(a, b) {
+  return a + b;
+}
+module.exports = sum;
