+++ sum.test.js
@@ -1,0 +1,5 @@
+const sum = require('./sum.js');
+
+test('adds 1 + 2 to equal 3', () => {
+  expect(sum(1, 2)).toBe(3);
+});
